requirements.txt
discord.py
python-telegram-bot
python-dotenv
colorama